refactor(kernel-timer): unify json formatting between reader and shared library
