+++ profiling/simple-kernel-timer/CMakeLists.txt
@@ -16,4 +16,4 @@ kp_add_executable(kp_reader kp_reader.cpp)
 target_link_libraries(kp_reader PRIVATE kp_kernel_timer)
 
 kp_add_executable(kp_json_writer kp_json_writer.cpp)
-target_link_libraries(kp_json_writer PRIVATE kp_kernel_timer)
+target_link_libraries(kp_json_writer PRIVATE kp_kernel_timer kp_kernel_shared)
